v2.0.0 发布

+++ src/main/java/com/ofwise/util/query/OffsetLimitRequest.java
@@ -134,7 +134,7 @@ public class OffsetLimitRequest implements Pageable, Serializable {
 
     @Override
     public Pageable withPage(int pageNumber) {
-        return new OffsetLimitRequest(getPageSize() * pageNumber, getPageSize(), getSort());
+        return new OffsetLimitRequest(getPageSize() * (long) pageNumber, getPageSize(), getSort());
     }
 
     @Override
